+++ rsa.h
@@ -5,8 +5,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-int is_prime(long int);
-int is_even(long int);
-void find_factors(long int);
+int is_prime(unsigned long long int);
+int is_even(unsigned long long int);
+void find_factors(unsigned long long int);
 
 #endif
